chore: merge CustomMetricButtons with FieldButtons
